Crate ocl

| GitHub

Rust implementation of OpenCL™.

This documentation is occasionally built from development branches and may differ slightly from what is on crates.io and the master branch.

The pages within are very much a work in progress, as is the library itself. Please help by filing an issue about unclear and/or incomplete documentation and it will be addressed.

An explanation/tutorial of how dimensions and sizes of buffers and work queues are used will be coming as soon. Until then please see the examples.

Low Level Interfaces

For lower level interfaces and to use OpenCL features that have not yet been implemented on the standard (high-level) interface types, see the ocl-core and cl-sys crates.

Help Wanted

Please help complete any functionality you may need by filing an issue or creating a pull request.

Feedback appreciated

Suggestions and nitpicks are most welcome. This isn't ever going to be a busy repo so don't hesitate to file an issue to offer constructive criticism.

“OpenCL and the OpenCL logo are trademarks of Apple Inc. used by permission by Khronos.”

Reexports

pub extern crate ocl_core as core;

Modules

aliases
Type aliases and structs meant to mirror those available within a kernel.

builders
Builders and associated settings-related types.

enums
Enumerators for settings and information requests.

ffi
FFI bindings for OpenCL.

flags
Bitflags for various parameter types.

traits
Commonly used traits.

util
Utility and debugging functions.

Structs

Buffer
A chunk of memory physically located on a device, such as a GPU.

Context
A context for a particular platform and set of device types.

Device
An individual device identifier (an OpenCL device_id).

Event
An event representing a command or user created event.

EventList
A list of events for coordinating enqueued commands.

Image
A section of device memory which represents one or many images.

Kernel
A kernel which represents a 'procedure'.

Platform
A platform identifier.

ProQue
An all-in-one chimera of the Program, Queue, Context and (optionally) SpatialDims types.

Program
A program from which kernels can be created from.

Queue
A command queue which manages all actions taken on kernels, buffers, and images.

Sampler
An image sampler used to process images.

Enums

Error
An enum containing either a String or one of several other error types.

SpatialDims
Specifies a size or offset in up to three dimensions.

Type Definitions

Result
ocl::Error result type.