Struct ocl::async::FutureGuard

#[must_use = "futures do nothing unless polled"]

pub struct FutureGuard<V, G> where
    G: OrderGuard<V>,  { /* fields omitted */ }

A future that resolves to a read or write guard after ensuring that the data being guarded is appropriately locked during the execution of an OpenCL command.

Waits until both an exclusive data lock can be obtained and all prerequisite OpenCL commands have completed.

Triggers an OpenCL command, remaining locked while the command executes.

Returns a guard which provides exclusive (write) or shared (read) access to the locked data.

Methods

impl<V, G> FutureGuard<V, G> where
G: OrderGuard<V>,

pub fn set_lock_wait_events<L: Into<EventList>>(&mut self, wait_events: L)

Sets an event wait list.

Setting a wait list will cause this FutureGuard to wait until contained events have their status set to complete before obtaining a lock on the guarded internal value.

[UNSTABLE]: This method may be renamed or otherwise changed at any time.

pub fn ewait_lock<L: Into<EventList>>(self, wait_events: L) -> FutureGuard<V, G>

Sets an event wait list.

See ::set_lock_wait_events.

[UNSTABLE]: This method may be renamed or otherwise changed at any time.

pub fn create_lock_event<C: ClContextPtr>(
&mut self,
context: C
) -> OclResult<&Event>

Creates an event which will be triggered when a lock is obtained on the guarded internal value.

The returned event can be added to the wait list of subsequent OpenCL commands with the expectation that when all preceding futures are complete, the event will automatically be 'triggered' by having its status set to complete, causing those commands to execute. This can be used to inject host side code in amongst OpenCL commands without thread blocking or extra delays of any kind.

pub fn enew_lock<C, En>(self, context: C, enew: En) -> FutureGuard<V, G> where
C: ClContextPtr,
En: ClNullEventPtr,

Creates an event which will be triggered when a lock is obtained on the guarded internal value.

enew must be an empty (null) event or event list.

See ::create_lock_event

Panics

Panics if there is an error creating the lock event.

[UNSTABLE]: This method may be renamed or otherwise changed at any time.

pub fn set_command_wait_event(&mut self, command_event: Event)

Sets a command completion wait event.

command_event must be an event created by enqueuing an OpenCL command which interacts (reads/writes) with the data associated with this FutureGuard.

If the command completion event is specified, this FutureGuard will "suffix" itself with an additional future that will wait until the command completes before resolving.

Not specifying a command completion event will cause this FutureGuard to resolve into an OrderGuard immediately after the lock is obtained (indicated by the optionally created lock event).

[UNSTABLE]: This method may be renamed or otherwise changed at any time.

pub fn ewait_command(self, command_event: Event) -> FutureGuard<V, G>

Sets a command completion wait event.

See ::set_command_wait_event.

[UNSTABLE]: This method may be renamed or otherwise changed at any time.

pub fn create_release_event<C: ClContextPtr>(
&mut self,
context: C
) -> OclResult<&Event>

Creates an event which will be triggered after this future resolves and the ensuing OrderGuard is dropped or manually released.

The returned event can be added to the wait list of subsequent OpenCL commands with the expectation that when all preceding futures are complete, the event will automatically be 'triggered' by having its status set to complete, causing those commands to execute. This can be used to inject host side code in amongst OpenCL commands without thread blocking or extra delays of any kind.

pub fn enew_release<C, En>(self, context: C, enew: En) -> FutureGuard<V, G> where
C: ClContextPtr,
En: ClNullEventPtr,

Creates an event which will be triggered after this future resolves and the ensuing OrderGuard is dropped or manually released.

enew must be an empty (null) event or event list.

See ::create_release_event.

Panics

Panics if there is an error creating the release event.

[UNSTABLE]: This method may be renamed or otherwise changed at any time.

pub fn lock_event(&self) -> Option<&Event>

Returns a reference to the event previously created with ::create_lock_event or ::enew_lock which will trigger (be completed) when the wait events are complete and the lock is locked.

pub fn release_event(&self) -> Option<&Event>

Returns a reference to the event previously created with ::create_release_event or ::enew_release which will trigger (be completed) when a lock is obtained on the guarded internal value.

pub fn wait(self) -> OclResult<G>

Blocks the current thread until the OpenCL command is complete and an appropriate lock can be obtained on the underlying data.

pub fn as_ptr(&self) -> *const V

Returns a mutable pointer to the data contained within the internal value, bypassing all locks and protections.

Panics

This future must not have already resolved into a guard.

pub fn as_mut_ptr(&self) -> *mut V

Returns a mutable pointer to the data contained within the internal value, bypassing all locks and protections.

Panics

This future must not have already resolved into a guard.

pub fn order_lock(&self) -> &OrderLock<V>

Returns a reference to the OrderLock used to create this future.

impl<V> FutureGuard<V, ReadGuard<V>>

pub fn upgrade_after_command(self) -> FutureGuard<V, WriteGuard<V>>