Module ocl::buffer

Interfaces to an OpenCL buffer.

Modules

tests

Structs

Buffer
An OpenCL buffer with an optional vector.

Functions

scrambled_vec
Returns a vector with length size containing random values in the (half-open) range [min_val, max_val).

shuffle_vec
Shuffles the values in a vector using a single pass of Fisher-Yates with a weak (not cryptographically secure) random number generator.

shuffled_vec
Returns a vector with length size which is first filled with each integer value in the (inclusive) range [min_val, max_val]. If size is greater than the number of integers in the aforementioned range, the integers will repeat. After being filled with size values, the vector is shuffled and the order of its values is randomized.