Struct ocl::buffer::Buffer

pub struct Buffer<T> {
    // some fields omitted
}

An OpenCL buffer with an optional vector.

One copy of data is stored in a buffer on the device associated with queue. A second is stored as an optional vector (vec) in host memory. Basically just a buffer with an optional built-in vector for use as a workspace etc.

The host side vector must be manually synchronized with the device side buffer using the .fill_vec(), .flush_vec(), etc. Data within the contained vector should generally be considered stale except immediately after a fill/flush (exception: pinned memory).

Fill/flush functions are for convenience and are equivalent to the psuedocode: read_into(self.vec) and write_from(self.vec).

Methods

impl<T: OclNum> Buffer<T>

Panics

All functions will panic upon any OpenCL error.

fn new<E: BufferDims>(dims: E, queue: &Queue) -> Buffer<T>

Creates a new read/write Buffer with dimensions: dims which will use the command queue: queue (and its associated device and context) for all operations.

The device side buffer will be allocated a size based on the maximum workgroup size of the device. This helps ensure that kernels do not attempt to read from or write to memory beyond the length of the buffer (see crate level documentation for more details about how dimensions are used). The buffer will be initialized with a sensible default value (probably 0).

Panics

The returned Buffer contains no host side vector. Functions associated with one such as .flush_vec(), fill_vec(), etc. will panic.

fn with_vec<E: BufferDims>(dims: E, queue: &Queue) -> Buffer<T>

Creates a new read/write Buffer with a host side working copy of data. Host vector and device buffer are initialized with a sensible default value.

fn with_vec_initialized_to<E: BufferDims>(init_val: T, dims: E, queue: &Queue) -> Buffer<T>

[UNSTABLE][MARKED FOR POSSIBLE REMOVAL]: Convenience function. Creates a new read/write Buffer with a host side working copy of data. Host vector and device buffer are initialized with the value, init_val.

fn with_vec_shuffled<E: BufferDims>(min_val: T, max_val: T, dims: E, queue: &Queue) -> Buffer<T>

[UNSTABLE][MARKED FOR POSSIBLE REMOVAL]: Convenience function. Creates a new read/write Buffer with a vector initialized with a series of integers ranging from min_val to max_val (closed) which are shuffled randomly.

Note: Even if the Buffer type is a floating point type, the values returned will still be integral values (e.g.: 1.0, 2.0, 3.0, etc.).

Security

Resulting values are not cryptographically secure.

fn with_vec_scrambled<E: BufferDims>(min_val: T, max_val: T, dims: E, queue: &Queue) -> Buffer<T>

[UNSTABLE][MARKED FOR POSSIBLE REMOVAL]: Convenience function. Creates a new read/write Buffer with a vector initialized with random values within the (half-open) range min_val..max_val.

Security

Resulting values are not cryptographically secure.

unsafe fn new_raw_unchecked(flags: cl_bitfield, len: usize, host_ptr: Option<&[T]>, queue: &Queue) -> Buffer<T>

[UNSTABLE] Creates a new Buffer with caller-managed buffer length, type, flags, and initialization.

Examples

See examples/buffer_unchecked.rs.

Parameter Reference Documentation

Refer to the following page for information about how to configure flags and other options for the buffer. https://www.khronos.org/registry/cl/sdk/1.2/docs/man/xhtml/clCreateBuffer.html

Safety

No creation time checks are made to help prevent device buffer overruns, etc. Caller should be sure to initialize the device side memory with a write. Any host side memory pointed to by host_ptr is at even greater risk if used with CL_MEM_USE_HOST_PTR (see below).

[IMPORTANT] Practically every read and write to an Buffer created in this way is potentially unsafe. Because .read() and .write() do not require an unsafe block, their implied promises about safety may be broken at any time.

You need to know what you're doing and be extra careful using an Buffer created with this function because it badly breaks Rust's usual safety promises even outside of an unsafe block.

This is horribly un-idiomatic Rust. You have been warned.

NOTE: The above important warnings probably only apply to buffers created with the CL_MEM_USE_HOST_PTR flag because the memory is considered 'pinned' but there may also be implementation specific issues which haven't been considered or are unknown.

Stability

The functionality of an Buffer used in this way may eventually be moved into a separate type, potentially called just Buffer which can manage its inherent unsafety without breaking promises. [update: probably not]

fn write(&mut self, data: &[T], offset: usize, wait_list: Option<&EventList>, dest_list: Option<&mut EventList>)

After waiting on events in wait_list to finish, writes the contents of 'data' to the remote device data buffer with a remote offset of offset and adds a new event to dest_list

If the dest_list event list is None, the write will be blocking.

fn flush_vec(&mut self, wait_list: Option<&EventList>, dest_list: Option<&mut EventList>)

After waiting on events in wait_list to finish, writes the contents of 'self.vec' to the remote device data buffer and adds a new event to dest_list

Panics

Panics if this Buffer contains no vector.

fn flush_vec_wait(&mut self)

Writes the contents of self.vec to the remote device data buffer and blocks until completed.

Equivalent to .flush_vec(None, None).

Panics

Panics if this Buffer contains no vector.

fn read(&self, data: &mut [T], offset: usize, wait_list: Option<&EventList>, dest_list: Option<&mut EventList>)

After waiting on events in wait_list to finish, reads the remote device data buffer with a remote offset of offset into data and adds a new event to dest_list.

If the dest_list event list is None, the read will be blocking.

fn fill_vec(&mut self, wait_list: Option<&EventList>, dest_list: Option<&mut EventList>)

After waiting on events in wait_list to finish, reads the remote device data buffer into 'self.vec' and adds a new event to dest_list.

Panics

Panics if this Buffer contains no vector.

fn fill_vec_wait(&mut self)

Reads the remote device data buffer into self.vec and blocks until completed.

Equivalent to .fill_vec(None, None).

Panics

Panics if this Buffer contains no vector.

fn wait(&self)

Blocks until the underlying command queue has completed all commands.

fn set_all_to(&mut self, val: T)

[MARKED FOR POSSIBLE REMOVAL]: Convenience function.

Panics

Panics if this Buffer contains no vector. [FIXME]: GET WORKING EVEN WITH NO CONTAINED VECTOR

fn set_range_to(&mut self, val: T, range: Range<usize>)

[MARKED FOR POSSIBLE REMOVAL]: Convenience function.

Panics

Panics if this Buffer contains no vector. [FIXME]: GET WORKING EVEN WITH NO CONTAINED VECTOR

fn len(&self) -> usize

Returns the length of the Buffer.

This is the length of both the device side buffer and the host side vector, if any. This may not agree with desired dataset size because it will have been rounded up to the nearest maximum workgroup size of the device on which it was created.

fn print_simple(&mut self)

[MARKED FOR POSSIBLE REMOVAL]: Convenience function. (move to tests module?)

Panics

Panics if this Buffer contains no vector. [FIXME]: GET WORKING EVEN WITH NO CONTAINED VECTOR

fn print_val_range(&mut self, every: usize, val_range: Option<(T, T)>)

[MARKED FOR POSSIBLE REMOVAL]: Convenience function. (move to tests module?)

Panics

Panics if this Buffer contains no vector. [FIXME]: GET WORKING EVEN WITH NO CONTAINED VECTOR

fn print(&mut self, every: usize, val_range: Option<(T, T)>, idx_range_opt: Option<Range<usize>>, zeros: bool)

[MARKED FOR POSSIBLE REMOVAL]: Convenience function. (move to tests module?) [FIXME]: CREATE A VECTOR FOR PRINTING IF NONE EXISTS

Panics

Panics if this Buffer contains no vector. [FIXME]: GET WORKING EVEN WITH NO CONTAINED VECTOR

unsafe fn resize(&mut self, new_dims: &BufferDims)

Resizes Buffer. Recreates device side buffer and dangles any references kernels may have had to the old buffer.

Safety

[IMPORTANT]: You must manually reassign any kernel arguments which may have had a reference to the (device side) buffer associated with this Buffer.

fn release(&mut self)

Decrements the reference count associated with the previous buffer object, self.buffer_obj.

fn vec(&self) -> OclResult<&Vec<T>>

Returns a reference to the local vector associated with this buffer.

Failures

Returns an error if this buffer contains no vector.

unsafe fn vec_mut(&mut self) -> OclResult<&mut Vec<T>>

Returns a mutable reference to the local vector associated with this buffer.

Failures

Returns an error if this buffer contains no vector.

Safety

Could cause data collisions, etc. Probably not unsafe strictly speaking (is it?) but marked as such to alert the caller to any potential synchronization issues.

unsafe fn get_unchecked(&self, idx: usize) -> &T

Returns an immutable reference to the value located at index idx, bypassing bounds and enum variant checks.

Safety

Assumes self.vec is a VecOption::Vec(_) and that the index idx is within the vector bounds.

unsafe fn get_unchecked_mut(&mut self, idx: usize) -> &mut T

Returns a mutable reference to the value located at index idx, bypassing bounds and enum variant checks.

Safety

Assumes self.vec is a VecOption::Vec(_) and that the index idx is within bounds. Might eat all the laundry.

fn buffer_obj(&self) -> cl_mem

Returns a copy of the device buffer object reference.

fn queue(&self) -> &Queue

Returns a reference to the program/command queue associated with this buffer.

unsafe fn set_queue(&mut self, queue: Queue)

Changes the queue used by this Buffer for reads and writes, etc.

Safety

Not all implications of changing the queue, particularly if the new queue is associated with a new device which has different workgroup size dimensions, have been considered or are dealt with. For now, considering these cases is left to the caller. It's probably a good idea to at least call .resize() after calling this function.

fn iter<'a>(&'a self) -> Iter<'a, T>

Returns an iterator to a contained vector.

Panics

Panics if this Buffer contains no vector.

fn iter_mut<'a>(&'a mut self) -> IterMut<'a, T>

Returns a mutable iterator to a contained vector.

Panics

Panics if this Buffer contains no vector.