Enum ocl::builders::ClWaitListPtrEnum

pub enum ClWaitListPtrEnum<'a> {
    Null,
    RawEventArray(&'a RawEventArray),
    EventCoreOwned(EventCore),
    EventOwned(Event),
    EventCore(&'a EventCore),
    Event(&'a Event),
    EventList(&'a EventList),
    EventSlice(&'a [Event]),
    EventPtrSlice(&'a [cl_event]),
    RefEventList(Ref<'a, EventList>),
    RefTraitObj(Ref<'a, ClWaitListPtr>),
    BoxTraitObj(Box<ClWaitListPtr>),
}

An enum which can represent several different ways of representing a event wait list.

Variants
NullRawEventArray(&'a RawEventArray)EventCoreOwned(EventCore)EventOwned(Event)EventCore(&'a EventCore)Event(&'a Event)EventList(&'a EventList)EventSlice(&'a [Event])EventPtrSlice(&'a [cl_event])RefEventList(Ref<'a, EventList>)RefTraitObj(Ref<'a, ClWaitListPtr>)BoxTraitObj(Box<ClWaitListPtr>)
Methods

impl<'a> ClWaitListPtrEnum<'a>

fn into_marker(self, queue: &Queue) -> OclResult<Event>

Converts this ClWaitListPtrEnum into a single marker event.

fn to_list(&self) -> EventList

Returns an EventList containing owned copies of each element in this ClWaitListPtrEnum.