Struct ocl::builders::KernelCmd

#[must_use = "commands do nothing unless enqueued"]

pub struct KernelCmd<'k> { /* fields omitted */ }

A kernel command builder used to queue a kernel with a mix of default and optionally specified arguments.

Methods

impl<'k> KernelCmd<'k>

A kernel enqueue command.

[UNSTABLE]: Methods still being tuned.

fn queue<'q, Q>(self, queue: &'q Q) -> KernelCmd<'k> where
'q: 'k,
Q: 'k + AsRef<CommandQueueCore>,

Specifies a queue to use for this call only.

fn gwo<D: Into<SpatialDims>>(self, gwo: D) -> KernelCmd<'k>

Specifies a global work offset for this call only.

fn gws<D: Into<SpatialDims>>(self, gws: D) -> KernelCmd<'k>

Specifies a global work size for this call only.

fn lws<D: Into<SpatialDims>>(self, lws: D) -> KernelCmd<'k>

Specifies a local work size for this call only.

fn ewait<'e, Ewl>(self, ewait: Ewl) -> KernelCmd<'k> where
'e: 'k,
Ewl: Into<ClWaitListPtrEnum<'e>>,

Specifies a list of events to wait on before the command will run.

fn ewait_opt<'e, Ewl>(self, ewait: Option<Ewl>) -> KernelCmd<'k> where
'e: 'k,
Ewl: Into<ClWaitListPtrEnum<'e>>,

Specifies a list of events to wait on before the command will run or resets it to None.

fn enew<'e, En>(self, new_event_dest: En) -> KernelCmd<'k> where
'e: 'k,
En: Into<ClNullEventPtrEnum<'e>>,

Specifies the destination list or empty event for a new, optionally created event associated with this command.

fn enew_opt<'e, En>(self, new_event_list: Option<En>) -> KernelCmd<'k> where
'e: 'k,
En: Into<ClNullEventPtrEnum<'e>>,

Specifies a destination list for a new, optionally created event associated with this command.

fn enq(self) -> OclResult<()>

Enqueues this kernel command.