Enum ocl::BuildOpt

pub enum BuildOpt {
    CmplrDefine {
        ident: String,
        val: String,
    },
    CmplrInclDir {
        path: String,
    },
    CmplrOther(String),
    IncludeDefine {
        ident: String,
        val: String,
    },
    IncludeRaw(String),
}

A build option intended for use either by the compiler as a command line switch or for inclusion at the top of the final build source code.

A few of the often used variants have constructors for convenience. [FIXME] TODO: Explain how each variant is used.

Variants

CmplrDefine
Fields

ident

val

CmplrInclDir
Fields

path

CmplrOther

IncludeDefine
Fields

ident

val

IncludeRaw

Methods

impl BuildOpt

fn cmplr_def<S: Into<String>>(ident: S, val: i32) -> BuildOpt

Returns a BuildOpt::CmplrDefine.

fn cmplr_opt<S: Into<String>>(opt: S) -> BuildOpt

Returns a BuildOpt::CmplrOther.

fn include_def<S: Into<String>>(ident: S, val: String) -> BuildOpt

Returns a BuildOpt::IncludeDefine.