Enum ocl::Error

pub enum Error {
    Void,
    Conversion(String),
    Status {
        status: Status,
        status_string: String,
        fn_name: &'static str,
        fn_info: String,
        desc: String,
    },
    String(String),
    Nul(NulError),
    Io(Error),
    FromUtf8Error(FromUtf8Error),
    UnspecifiedDimensions,
    IntoStringError(IntoStringError),
    EmptyInfoResult(EmptyInfoResult),
}

An enum containing either a String or one of several other error types.

Implements the usual error traits.

Stability

The String variant may eventually be removed. Many more variants and sub-types will be added as time goes on and things stabilize.

Status will eventually be changed internally to contain a sub-error type unique to each function which generates it (yeah that'll be fun to implement).

UnspecifiedDimensions may be moved into a sub-type.

For now, don't assume the existence of or check for any of the above.

Variants
VoidConversion(String)Status
Fields of Status

status: Status

status_string: String

fn_name: &'static str

fn_info: String

desc: String

String(String)Nul(NulError)Io(Error)FromUtf8Error(FromUtf8Error)UnspecifiedDimensionsIntoStringError(IntoStringError)EmptyInfoResult(EmptyInfoResult)
Methods

impl Error

fn new<S>(desc: S) -> Error where
S: Into<String>,

Deprecated since 0.4.0
: Use ::from instead.

Returns a new Error with the description string: desc.

Depricated

Use ::from instead.

fn string<S>(desc: S) -> Error where
S: Into<String>,

Deprecated since 0.4.0
: Use ::from instead.

Returns a new Error::String with the given description.

fn err<T, S>(desc: S) -> Result<T, Error> where
S: Into<String>,

Deprecated since 0.4.0
: Use Err("...".into()) instead.

Returns a new ocl_core::Result::Err containing an ocl_core::Error::String variant with the given description.

Depricated

Use ::err_string or Err("...".into()) instead.

fn err_string<T, S>(desc: S) -> Result<T, Error> where
S: Into<String>,

Returns a new Err(ocl_core::Error::String(...)) variant with the given description.

fn eval_errcode<T, S>(
errcode: i32,
result: T,
fn_name: &'static str,
fn_info: S
) -> Result<T, Error> where
S: Into<String>,

Returns a new ocl::Result::Err containing an ocl::Error with the given error code and description.

fn err_conversion<T, S>(desc: S) -> Result<T, Error> where
S: Into<String>,

Returns a new ocl::Result::Err containing an ocl::Error::Conversion variant with the given description.

fn prepend<'s, S>(&'s mut self, txt: S) where
S: AsRef<&'s str>,

If this is a String variant, concatenate txt to the front of the contained string. Otherwise, do nothing at all.

fn status(&self) -> Option<Status>

Returns the error status code for Status variants.