Enum ocl::SimpleDims

pub enum SimpleDims {
    Unspecified,
    OneDim(usize),
    TwoDims(usize, usize),
    ThreeDims(usize, usize, usize),
}

A simple implementation of a type specifying the sizes of up to three dimensions.

Custom types implementing BufferDims can and should be created to express more complex relationships between buffer and work size.

[FIXME] TODO: Much more explaination needed as soon as conventions solidify. [UNSTABLE]: MAY BE CONSOLIDATED WITH WorkSize.

Variants

Unspecified

OneDim

TwoDims

ThreeDims

Methods

impl SimpleDims

fn new(d0: Option<usize>, d1: Option<usize>, d2: Option<usize>) -> Result<SimpleDims, OclError>

Returns a new SimpleDims.

Dimensions must be specified in order from d0 -> d1 -> d2; i.e. d1 cannot be Some(x) if d0 is None.

fn work_size(&self) -> WorkSize

Returns a WorkSize corresponding to the dimensions of this SimpleDims.