Enum ocl::enums::ImageChannelDataType

pub enum ImageChannelDataType {
    SnormInt8,
    SnormInt16,
    UnormInt8,
    UnormInt16,
    UnormShort565,
    UnormShort555,
    UnormInt101010,
    SignedInt8,
    SignedInt16,
    SignedInt32,
    UnsignedInt8,
    UnsignedInt16,
    UnsignedInt32,
    HalfFloat,
    Float,
    UnormInt24,
}

Describes the size of the channel data type. The number of bits per element determined by the image_channel_data_type and image_channel_order must be a power of two. The list of supported values is described in the table below. (from SDK)

Variants
SnormInt8SnormInt16UnormInt8UnormInt16UnormShort565UnormShort555UnormInt101010SignedInt8SignedInt16SignedInt32UnsignedInt8UnsignedInt16UnsignedInt32HalfFloatFloatUnormInt24