Struct ocl::BuildConfig

pub struct BuildConfig {
    // some fields omitted
}

Compilation options for building an OpenCL program. Used when creating a new Program.

Methods

impl BuildConfig

fn new() -> BuildConfig

Returns a new, empty, build configuration object.

fn cmplr_def(self, name: &'static str, val: i32) -> BuildConfig

Adds a compiler command line definition => -D {name}={val} (builder-style).

fn cmplr_opt(self, co: &'static str) -> BuildConfig

Adds a raw compiler command line option => {co} (builder-style).

fn bo(self, bo: BuildOpt) -> BuildConfig

Pushes pre-created build option to the list (builder-style).

fn kern_file(self, file_name: &'static str) -> BuildConfig

Adds a kernel file to the list of included sources (builder-style).

fn kern_embed(self, source: &'static str) -> BuildConfig

Adds text to the included kernel source (builder-style).

fn add_kern_file(&mut self, file_name: String)

Adds a kernel file to the list of included sources.

fn add_embedded_kern(&mut self, source: String)

Adds text to the included kernel source.

fn kernel_file_names(&self) -> &Vec<String>

Returns a list of kernel file names added for inclusion in the build.

fn compiler_options(&self) -> Result<CString, NulError>

Returns a contatenated string of compiler command line options used when building a Program.

fn kernel_includes(&self) -> Result<Vec<CString>, NulError>

Parses self.options for options intended for inclusion at the beginning of the final program source and returns them as a list of strings.

Generally used for #define directives, constants, etc. Normally called from ::kernel_strings() but can also be called from anywhere for debugging purposes.

fn kernel_strings(&self) -> Result<Vec<CString>, String>

Returns the final program source code as a list of strings.

Order of inclusion: - includes from ::kernel_includes() - source from files listed in self.kernel_file_names in reverse order - raw source from self.embedded_kernel_source [UNSTABLE] TODO: Fix up error handling: return an OclResult.