Struct ocl::Event

#[repr(C)]
#[must_use = "futures do nothing unless polled"]

pub struct Event(_);

An event representing a command or user created event.

Methods

impl Event

fn empty() -> Event

Creates a new, empty (null) event which must be filled by a command, associating the event with it.

fn user<C: ClContextPtr>(context: C) -> OclResult<Event>

Creates a new, empty event which must be filled by a newly initiated command, associating the event with it.

fn is_empty(&self) -> bool

Returns true if this event is 'empty' and has not yet been associated with a command.

fn set_unpark_callback(&self) -> OclResult<()>

Sets a callback function to trigger upon completion of this event which will unpark the current task.

To be used within the context of a futures task.

Panics

This function will panic if a task is not currently being executed. That is, this method can be dangerous to call outside of an implementation of poll.

fn info(&self, info_kind: EventInfo) -> EventInfoResult

Returns info about the event.

fn profiling_info(&self, info_kind: ProfilingInfo) -> ProfilingInfoResult

Returns info about the event.

fn queue_core(&self) -> OclResult<CommandQueueCore>

Returns this event's associated command queue.

fn core(&self) -> &EventCore

Returns a reference to the core pointer wrapper, usable by functions in the core module.

fn into_raw(self) -> cl_event

Consumes the Event, returning the wrapped cl_event pointer.

To avoid a memory leak the pointer must be converted back to an Event using Event::from_raw.

unsafe fn from_raw(ptr: cl_event) -> Event

Constructs an Event from a raw cl_event pointer.

The raw pointer must have been previously returned by a call to a Event::into_raw.