Struct ocl::EventList

pub struct EventList { /* fields omitted */ }

A list of events for coordinating enqueued commands.

Events contain status information about the command that created them. Used to coordinate the activity of multiple commands with more fine-grained control than the queue alone.

For access to individual events use get_clone and last_clone then either store or discard the result.

Methods

impl EventList

fn new() -> EventList

Returns a new, empty, EventList.

fn push(&mut self, event: Event)

Adds an event to the list.

fn pop(&mut self) -> Option<Event>

Removes the last event from the list and returns it.

fn get_clone(&self, index: usize) -> Option<Event>

Returns a new copy of an event by index.

fn last_clone(&self) -> Option<Event>

Returns a copy of the last event in the list.

unsafe fn set_callback<T>(&self,
callback_receiver: Option<EventCallbackFn>,
user_data: &mut T)
-> OclResult<()>

Sets a callback function, callback_receiver, to trigger upon completion of the last event added to the event list with an optional reference to user data.

Safety

user_data must be guaranteed to still exist if and when callback_receiver is ever called.

TODO: Create a safer type wrapper for callback_receiver. TODO: Move this method to Event.

fn len(&self) -> usize

Returns the number of events in the list.

fn is_empty(&self) -> bool

Returns if there is no events.

fn core_as_ref(&self) -> &EventListCore

fn core_as_mut(&mut self) -> &mut EventListCore

fn wait(&self) -> OclResult<()>

Waits for all events in list to complete.