Struct ocl::ProQue

pub struct ProQue { /* fields omitted */ }

An all-in-one chimera of the Program, Queue, Context and (optionally) SpatialDims types.

Handy when you only need a single context, program, and queue for your project or when using a unique context and program on each device.

All ProQue functionality is also provided separately by the Context, Queue, Program, and SpatialDims types.

Creation

There are two ways to create a ProQue:

[Recommended] Use ProQue::builder or ProQueBuilder::new().

Call ::new and pass pre-created components.

Destruction

Now handled automatically. Freely use, store, clone, discard, share among threads... put some on your toast... whatever.

Methods

impl ProQue

fn builder() -> ProQueBuilder

Returns a new ProQueBuilder.

This is the recommended way to create a new ProQue.

Calling ProQueBuilder::build() will return a new ProQue.

fn new<D: Into<SpatialDims>>(
context: Context,
queue: Queue,
program: Program,
dims: Option<D>
) -> ProQue

Creates a new ProQue from individual parts.

Use ::builder instead unless you know what you're doing. Creating from components associated with different devices or contexts will cause errors later on.

fn create_kernel(&self, name: &str) -> OclResult<Kernel>

Creates a kernel with pre-assigned dimensions.

fn create_buffer<T: OclPrm>(&self) -> OclResult<Buffer<T>>

Returns a new buffer.

The default dimensions and queue from this ProQue will be used.

The buffer will be filled with zeros upon creation, blocking the current thread until completion.

Use Buffer::builder() (or BufferBuilder::new()) for access to the full range of buffer creation options.

Errors

This ProQue must have been pre-configured with default dimensions. If not, set them with ::set_dims, or just create a buffer using Buffer::builder() instead.

fn set_dims<S: Into<SpatialDims>>(&mut self, dims: S)

Sets the default dimensions used when creating buffers and kernels.

fn max_wg_size(&self) -> OclResult<usize>

Returns the maximum workgroup size supported by the device associated with this ProQue.

[UNSTABLE]: Evaluate usefulness.

fn queue(&self) -> &Queue

Returns a reference to the queue associated with this ProQue.

fn context(&self) -> &Context

Returns the contained context.

fn program(&self) -> &Program

Returns the current program build.

fn dims(&self) -> &SpatialDims

Returns the current dims or panics.

[UNSTABLE]: Evaluate which 'dims' method to keep. Leaning towards this version at the moment.

fn dims_result(&self) -> OclResult<&SpatialDims>

Returns the current dims or an error.

[UNSTABLE]: Evaluate which 'dims' method to keep. Leaning towards the above, panicing version at the moment.