Struct ocl::Program

pub struct Program { /* fields omitted */ }

A program from which kernels can be created from.

To use with multiple devices, create manually with ::from_parts().

Destruction

Handled automatically. Feel free to store, clone, and share among threads as you please.

Methods

impl Program

fn builder() -> ProgramBuilder

Returns a new ProgramBuilder.

fn new(src_strings: Vec<CString>,
cmplr_opts: CString,
context_obj_core: &ContextCore,
device_ids: &[Device])
-> OclResult<Program>

Returns a new program built from pre-created build components and device list.

Prefer ::builder to create a new Program.

fn core_as_ref(&self) -> &ProgramCore

Returns a reference to the core pointer wrapper, usable by functions in the core module.

fn devices(&self) -> &[Device]

Returns the list of devices associated with this program.

fn info(&self, info_kind: ProgramInfo) -> ProgramInfoResult

Returns info about this program.

fn build_info(&self,
device: Device,
info_kind: ProgramBuildInfo)
-> ProgramBuildInfoResult

Returns info about this program's build.

TODO: Check that device is valid.