Struct ocl::Queue

pub struct Queue { /* fields omitted */ }

A command queue which manages all actions taken on kernels, buffers, and images.

Methods

impl Queue

fn new(
context: &Context,
device: Device,
properties: Option<CommandQueueProperties>
) -> OclResult<Queue>

Returns a new Queue on the device specified by device.

fn flush(&self) -> OclResult<()>

Issues all previously queued OpenCL commands to the device.

fn finish(&self) -> OclResult<()>

Blocks until all commands in this queue have completed before returning.

fn enqueue_marker<Ewl>(&self, ewait: Option<Ewl>) -> OclResult<Event> where
Ewl: ClWaitListPtr,

Enqueues a marker command which waits for either a list of events to complete, or all previously enqueued commands to complete.

fn core_as_ref(&self) -> &CommandQueueCore

Deprecated since 0.13.0
: Use ::core instead.

Returns a reference to the core pointer wrapper, usable by functions in the core module.

fn core(&self) -> &CommandQueueCore

Returns a reference to the core pointer wrapper, usable by functions in the core module.

fn context(&self) -> Context

Returns a copy of the Context associated with this queue.

fn device(&self) -> Device

Returns the OpenCL device associated with this queue.

fn device_version(&self) -> OpenclVersion

Returns the cached device version.

fn info(&self, info_kind: CommandQueueInfo) -> CommandQueueInfoResult

Returns info about this queue.