Trait ocl::traits::OclPrm

pub unsafe trait OclPrm: 'static + Send + Copy + Sync + PartialEq<Self> + Clone + Default + Debug { }

A primitive type usable within OpenCL kernels.

Includes all of the signed, unsigned, and floating point 8 bit - 64 bit scalar primitives (ex.: cl_char, cl_uint, cl_double) (exception: cl_half) and their vector counterparts (ex.: cl_int4, cl_float3, cl_short16);

Can also be implemented for custom types as long as layout and alignment are conserved between Rust and OpenCL (repr "C").