Crate ocl

Rust implementation of OpenCL.

This documentation is very much a work in progress and is roughly 60% complete. Please help by filing an issue about unclear and/or incomplete documentation and it will be addressed (hopefully) quickly.

Links

GitHub: https://github.com/cogciprocate/ocl

crates.io:

Reexports

pub use self::cl_h::{cl_float, cl_char, cl_uchar, cl_short, cl_ushort, cl_int, cl_uint, cl_long, cl_bitfield, CLStatus, CL_DEVICE_TYPE_DEFAULT, CL_DEVICE_TYPE_CPU, CL_DEVICE_TYPE_GPU, CL_DEVICE_TYPE_ACCELERATOR, CL_DEVICE_TYPE_ALL};

pub use self::formatting as fmt;

pub use self::buffer::Buffer;

pub use self::buffer::tests::BufferTest;

Modules

buffer
Interfaces to an OpenCL buffer.

cl_h
Rust definitions for the OpenCL ABI

formatting
Tools for rendering objects to the screen or as text.

Structs

BuildConfig
Compilation options for building an OpenCL program. Used when creating a new Program.

Context
An OpenCL context for a particular platform and set of device types.

EventList
A list of OpenCL events which contain status information about the command that created them. Used to coordinate the activity of multiple commands.

Kernel
An OpenCL kernel.

OclError
Error type containing a string.

ProQue
A convenience wrapper chimera of Program and Queue.

Program
An OpenCL program, sometimes referred to as a build.

Queue
An OpenCL command queue.

Enums

BuildOpt
A build option intended for use either by the compiler as a command line switch or for inclusion at the top of the final build source code.

SimpleDims
A simple implementation of a type specifying the sizes of up to three dimensions.

WorkSize
Defines the amount of work to be done by a kernel for each of up to three dimensions.

Traits

BufferDims
A type which has dimensional properties allowing it to be used to define the size of buffers and work sizes.

OclNum
A number compatible with OpenCL.

Functions

padded_len
Pads len to make it evenly divisible by incr.

Type Definitions

OclResult
OclError result type.