Module ocl::builders

Builders and associated settings-related types.

Structs

BufferBuilder
A buffer builder.

BufferCmd
A buffer command builder used to enqueue reads, writes, fills, and copies.

BufferMapCmd
A command builder used to enqueue a map command.

BufferReadCmd
A buffer command builder used to enqueue reads.

BufferWriteCmd
A buffer command builder used to enqueue writes.

ContextBuilder
A builder for Context.

ContextProperties
Context properties list.

ImageBuilder
A builder for Image.

ImageCmd
An image command builder for enqueuing reads, writes, fills, and copies.

ImageDescriptor
An image descriptor use in the creation of Image.

ImageFormat
Image format properties used by Image.

KernelBuilder
A kernel builder.

KernelCmd
A kernel command builder used to enqueue a kernel with a mix of default and optionally specified arguments.

ProQueBuilder
A builder for ProQue.

ProgramBuilder
A builder for Program.

Enums

BufferCmdDataShape
The 'shape' of the data to be processed, whether one or multi-dimensional.

BufferCmdKind
The type of operation to be performed by a command.

BuildOpt
A build option used by ProgramBuilder.

ClNullEventPtrEnum

ClWaitListPtrEnum
An enum which can represent several different ways of representing a event wait list.

DeviceSpecifier
Specifies [what boils down to] a list of devices.

ImageCmdKind
The type of operation to be performed by a command.