Enum ocl::SpatialDims

pub enum SpatialDims {
    Unspecified,
    One(usize),
    Two(usize, usize),
    Three(usize, usize, usize),
}

Specifies a size or offset in up to three dimensions.

Using SpatialDims to specify dimensions for your application may not be appropriate. Custom types implementing the traits MemLen and WorkDims should be created to express more complex relationships between data shape and work size for kernels which do not have a 1:1 correspondence between data set length and global work size.

[FIXME]: Describe the ways a SpatialDims can be created using various From implementations.

[UNSTABLE]: This type and its methods may be renamed or otherwise changed at any time. This is still a work in progress.

Variants
UnspecifiedOne(usize)Two(usize, usize)Three(usize, usize, usize)
Methods

impl SpatialDims

pub fn new(
d0: Option<usize>,
d1: Option<usize>,
d2: Option<usize>
) -> OclResult<SpatialDims>

Returns a new SpatialDims.

Dimensions must be specified in order from d0 -> d1 -> d2; i.e. d1 cannot be Some(x) if d0 is None.

Explicitly setting all zeros as values is not recommended as it is invalid to OpenCL functions. Use ::Unspecified to represent NULL instead.

pub fn dim_count(&self) -> u32

Returns the number of dimensions defined by this SpatialDims.

pub fn to_lens(&self) -> Result<[usize; 3], UnspecifiedDimensionsError>

Returns a 3D size or an error if unspecified.

pub fn to_offset(&self) -> Result<[usize; 3], UnspecifiedDimensionsError>

Returns a 3D offset or an error if unspecified.

pub fn to_len(&self) -> usize

Returns the product of all contained dimensional values (equivalent to a length, area, or volume) or zero if unspecified.

Check ::is_unspecified if it matters whether or not the zero length has been explicitly set however, explicitly setting all zeros as dimensions is not recommended as it is an invalid value to OpenCL functions. Use ::Unspecified to represent NULL instead.

pub fn try_to_padded_len(&self, incr: usize) -> OclResult<usize>

Takes the length and rounds it up to the nearest incr or an error.

pub fn is_unspecified(&self) -> bool

Returns true if this SpatialDims is an Unspecified variant.