Struct ocl::Device

#[repr(C)]

pub struct Device(_);

An individual device identifier (an OpenCL device_id).

Methods

impl Device

pub fn first<P: Borrow<Platform>>(platform: P) -> OclResult<Device>

Returns the first available device on a platform.

pub fn by_idx_wrap<P: Borrow<Platform>>(
platform: P,
device_idx_wrap: usize
) -> OclResult<Device>

Returns a single device specified by a wrapped index.

pub fn specifier() -> DeviceSpecifier

Returns a DeviceSpecifier useful for precisely specifying a set of devices.

pub fn resolve_idxs(
idxs: &[usize],
devices: &[Device]
) -> OclResult<Vec<Device>>

Resolves a list of indices into a list of valid devices.

devices is the set of all indexable devices.

Errors

All indices in idxs must be valid. Use resolve_idxs_wrap for index lists which may contain out of bounds indices.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn resolve_idxs_wrap(idxs: &[usize], devices: &[Device]) -> Vec<Device>

Resolves a list of indices into a list of valid devices.

devices is the set of all indexable devices.

Wraps indices around using modulo (%) so that every index is valid.

pub fn list<P: Borrow<Platform>>(
platform: P,
device_types: Option<DeviceType>
) -> OclCoreResult<Vec<Device>>

Returns a list of all devices avaliable for a given platform which optionally match the flags set in the bitfield, device_types.

Setting device_types to None will return a list of all avaliable devices for platform regardless of type.

Errors

Returns an Err(ocl::core::Error::Status {...}) enum variant upon any OpenCL error. Calling .status() on the returned error will return an Option(``[ocl::core::Status]``) which can be unwrapped then matched to determine the precise reason for failure.

pub fn list_all<P: Borrow<Platform>>(platform: P) -> OclCoreResult<Vec<Device>>

Returns a list of all devices avaliable for a given platform.

Equivalent to ::list(platform, None).

See ::list for other error information.

pub fn list_select<P: Borrow<Platform>>(
platform: P,
device_types: Option<DeviceType>,
idxs: &[usize]
) -> OclResult<Vec<Device>>

Returns a list of devices filtered by type then selected using a list of indices.

Errors

All indices in idxs must be valid.

See ::list for other error information.

pub fn list_select_wrap<P: Borrow<Platform>>(
platform: P,
device_types: Option<DeviceType>,
idxs: &[usize]
) -> OclCoreResult<Vec<Device>>

Returns a list of devices filtered by type then selected using a wrapping list of indices.

Wraps indices around (%) so that every index is valid.

Errors

See ::list

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn list_from_core(devices: Vec<DeviceIdCore>) -> Vec<Device>

Returns a list of Devices from a list of DeviceIdCores

pub fn name(&self) -> OclCoreResult<String>

Returns the device name.

pub fn vendor(&self) -> OclCoreResult<String>

Returns the device vendor as a string.

pub fn max_wg_size(&self) -> OclCoreResult<usize>

Returns the maximum workgroup size or an error.

pub fn mem_base_addr_align(&self) -> OclCoreResult<u32>

Returns the memory base address alignment offset or an error.

pub fn is_available(&self) -> OclCoreResult<bool>

Returns whether or not the device is available for use.

pub fn info(&self, info_kind: DeviceInfo) -> OclCoreResult<DeviceInfoResult>

Returns info about the device.

pub fn to_string(&self) -> String

Returns a string containing a formatted list of device properties.

pub fn as_core(&self) -> &DeviceIdCore

Returns the underlying DeviceIdCore.