Struct ocl::Event

pub struct Event(_);

An event representing a command or user created event.

Methods

impl Event

fn empty() -> Event

Creates a new, empty event which must be filled by a newly initiated command, becoming associated with it.

unsafe fn from_core(event_core: EventCore) -> Event

Creates a new Event from a EventCore.

Safety

Not meant to be called directly.

fn wait(&self) -> OclResult<()>

Waits for all events in list to complete before returning.

Similar in function to Queue::finish().

fn info(&self, info_kind: EventInfo) -> EventInfoResult

Returns info about the event.

fn profiling_info(&self, info_kind: ProfilingInfo) -> ProfilingInfoResult

Returns info about the event.

fn core_as_ref(&self) -> Option<&EventCore>

Returns a reference to the core pointer wrapper, usable by functions in the core module.

fn core_as_mut(&mut self) -> Option<&mut EventCore>

Returns a mutable reference to the core pointer wrapper usable by functions in the core module.

fn is_empty(&self) -> bool

Returns true if this event is 'empty' and has not yet been associated with a command.