Struct ocl::EventList

pub struct EventList {
    // some fields omitted
}

A list of OpenCL events which contain status information about the command that created them. Used to coordinate the activity of multiple commands.

Methods

impl EventList

fn new() -> EventList

Returns a new, empty, EventList.

fn union(&self, other_list: &EventList) -> EventList

Merges the copied contents of this list and another into a new list and returns it.

fn allot(&mut self) -> &mut [cl_event]

Appends a new null element to the end of the list and returns a mutable slice containing only that element.

fn events(&self) -> &[cl_event]

Returns an immutable slice to the events list.

fn as_ptr(&self) -> *const cl_event

Returns a const pointer to the list, useful for passing directly to the c ffi.

fn wait(&self)

Waits for all events in list to complete.

unsafe fn set_callback<T>(&self, callback_receiver: extern "C" fn(cl_event, cl_int, *mut c_void), user_data: &mut T)

Sets a callback function, callback_receiver, to trigger upon completion of the last event added to the event list with an optional reference to user data.

Safety

user_data must be guaranteed to still exist if and when callback_receiver is ever called.

fn count(&self) -> u32

Returns the number of events in the list.

fn clear(&mut self)

Clears this list regardless of whether or not its events have completed.

fn clear_completed(&mut self)

Clears each completed event from the list.

fn release_all(&mut self)

Releases all events in the list by decrementing their reference counts by one and empties the list.

Events will continue to exist until their creating commands have completed and no other commands are waiting for them to complete.