Struct ocl::EventList

pub struct EventList { /* fields omitted */ }

A list of events for coordinating enqueued commands.

Events contain status information about the command that created them. Used to coordinate the activity of multiple commands with more fine-grained control than the queue alone.

For access to individual events use get_clone and last_clone then either store or discard the result.

EventList is a dynamically allocated list. It will be (internally) stack allocated (as an [Event; 8]) until it reaches a length of 9 at which time it will become heap-allocated (a Vec<Event>).

At the time of this writing, converting back from heap to stack allocation is not implemented but a bit of prodding (by filing an issue) would probably work to remedy that if anyone thinks it useful.

Methods

impl EventList

fn new() -> EventList

Returns a new, empty, stack-allocated EventList.

fn with_capacity(cap: usize) -> EventList

Returns a new, empty, EventListwith an initial capacity ofcap`.

If cap is greater than 8, the event list will be heap-allocated.

fn push<E: Into<Event>>(&mut self, event: E)

Adds an event to the list.

fn push_some<E: Into<Event>>(&mut self, event: Option<E>)

Pushes an Option<Event> to the list if it is Some(...).

fn pop(&mut self) -> Option<Event>

Removes the last event from the list and returns it.

fn clear(&mut self)

Clears all events from the list whether or not they have completed.

Forwards any errors related to releasing events.

fn clear_completed(&mut self) -> OclResult<()>

Clears events which have completed.

fn wait_for(&self) -> OclResult<()>

Blocks the host thread until all events in this list are complete.

fn enqueue_marker(&self, queue: &Queue) -> OclResult<Event>

Enqueue a marker event representing the completion of each and every event in this list.

Platform Compatibility

Some device/platform combinations (particularly older Intel CPUs on Intel platform drivers) may have intermittent issues waiting on markers when multiple threads are in use. This is rare and can be circumvented by using AMD platform drivers instead. Please file an issue immediately if you run into problems on your platform so that we may make note of it here in the documentation.

fn as_slice(&self) -> &[Event]

Returns a slice of the contained events.

fn as_mut_slice(&mut self) -> &mut [Event]

Returns a mutable slice of the contained events.