Struct ocl::Kernel

pub struct Kernel {
    // some fields omitted
}

An OpenCL kernel.

[FIXME] TODO: Add more details, examples, etc. [FIXME] TODO: Add information about panics and errors.

Methods

impl Kernel

fn new(name: String, program: &Program, queue: &Queue, gws: WorkSize) -> Kernel

Returns a new kernel.

fn gwo(self, gwo: WorkSize) -> Kernel

Sets the global work offset (builder-style).

fn lws(self, lws: WorkSize) -> Kernel

Sets the local work size (builder-style).

fn arg_env<T: OclNum>(self, buffer: &Buffer<T>) -> Kernel

Adds a new argument to the kernel specifying the buffer object represented by 'buffer' (builder-style). Argument is added to the bottom of the argument order.

fn arg_scl<T: OclNum>(self, scalar: T) -> Kernel

Adds a new argument specifying the value: scalar (builder-style). Argument is added to the bottom of the argument order.

fn arg_loc<T: OclNum>(self, length: usize) -> Kernel

Adds a new argument specifying the allocation of a local variable of size length * sizeof(T) bytes (builder_style).

Local variables are used to share data between work items in the same workgroup.

fn arg_scl_named<T: OclNum>(self, name: &'static str, scalar_opt: Option<T>) -> Kernel

Adds a new named argument (in order) specifying the value: scalar (builder-style).

Named arguments can be easily modified later using ::set_arg_scl_named().

fn arg_env_named<T: OclNum>(self, name: &'static str, buffer_opt: Option<&Buffer<T>>) -> Kernel

Adds a new named buffer argument specifying the buffer object represented by 'buffer' (builder-style). Argument is added to the bottom of the argument order.

Named arguments can be easily modified later using ::set_arg_scl_named().

fn new_arg_buffer<T: OclNum>(&mut self, buffer_opt: Option<&Buffer<T>>) -> u32

Non-builder-style version of ::arg_env().

fn new_arg_scalar<T: OclNum>(&mut self, scalar_opt: Option<T>) -> u32

Non-builder-style version of ::arg_scl().

fn new_arg_local<T: OclNum>(&mut self, length: usize) -> u32

Non-builder-style version of ::arg_loc().

fn set_arg_scl_named<T: OclNum>(&mut self, name: &'static str, scalar: T)

Modifies the kernel argument named: name.

fn set_arg_env_named<T: OclNum>(&mut self, name: &'static str, buffer: &Buffer<T>)

Modifies the kernel argument named: name.

fn enqueue_with_cmd_queue(&self, cmd_queue: cl_command_queue, wait_list: Option<&EventList>, dest_list: Option<&mut EventList>)

Enqueues kernel using a non-default command queue.

The command queue, cmd_queue, must have the same context and device as the default command queue passed when creating kernel.

fn enqueue(&self, wait_list: Option<&EventList>, dest_list: Option<&mut EventList>)

Enqueues kernel on the default command queue.

fn arg_count(&self) -> u32

Returns the number of arguments specified for this kernel.