Struct ocl::Platform

#[repr(C)]

pub struct Platform(_);

A platform identifier.

Methods

impl Platform

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn list() -> Vec<Platform>

Returns a list of all platforms avaliable on the host machine.

pub fn first() -> OclResult<Platform>

Returns the first available platform.

This method differs from Platform::default() in two ways. First, it ignores the OCL_DEFAULT_PLATFORM_IDX environment variable (Platform::default always respects it). Second, this function will not panic if no platforms are available but will instead return an error.

pub fn new(id_core: PlatformIdCore) -> Platform

Creates a new Platform from a PlatformIdCore.

Safety

Not meant to be called unless you know what you're doing.

Use list to get a list of platforms.

Important traits for Vec<u8>
impl Write for Vec<u8>

pub fn list_from_core(platforms: Vec<PlatformIdCore>) -> Vec<Platform>

Returns a list of Platforms from a list of PlatformIdCores

pub fn info(&self, info_kind: PlatformInfo) -> OclResult<PlatformInfoResult>

Returns info about the platform.

pub fn profile(&self) -> OclResult<String>

Returns the platform profile as a string.

Returns the profile name supported by the implementation. The profile name returned can be one of the following strings:

FULL_PROFILE - if the implementation supports the OpenCL specification (functionality defined as part of the core specification and does not require any extensions to be supported).

EMBEDDED_PROFILE - if the implementation supports the OpenCL embedded profile. The embedded profile is defined to be a subset for each version of OpenCL.

pub fn version(&self) -> OclResult<String>

Returns the platform driver version as a string.

Returns the OpenCL version supported by the implementation. This version string has the following format:

OpenCL<major_version.minor_version><platform-specific information>

The major_version.minor_version value returned will be '1.2'.

TODO: Convert this to new version system returning an OpenclVersion.

pub fn name(&self) -> OclResult<String>

Returns the platform name as a string.

pub fn vendor(&self) -> OclResult<String>

Returns the platform vendor as a string.

pub fn extensions(&self) -> OclResult<Extensions>

Returns the list of platform extensions.

Extensions defined here must be supported by all devices associated with this platform.

pub fn as_core(&self) -> &PlatformIdCore

Returns a reference to the underlying PlatformIdCore.