Struct ocl::ProQue

pub struct ProQue {
    // some fields omitted
}

A convenience wrapper chimera of Program and Queue.

Useful when using a unique program build on each device.

Methods

impl ProQue

fn new(context: &Context, device_idx: Option<usize>) -> ProQue

Creates a new queue on the device with device_idx (see Queue documentation) and returns a new Program/Queue hybrid.

::build() must be called before this ProQue can be used.

fn build(&mut self, build_config: BuildConfig) -> Result<(), String>

Builds contained program with build_config.

Panics

This ProQue must contain a program.

fn clear_build(&mut self)

Clears the current program build.

fn create_kernel(&self, name: &str, gws: WorkSize) -> Kernel

Returns a new Kernel with name: name and global work size: gws.

fn get_max_work_group_size(&self) -> usize

Returns the maximum workgroup size supported by the device on which the contained queue exists.

fn queue(&self) -> &Queue

Returns the queue created when constructing this ProQue.

fn program(&self) -> &Option<Program>

Returns the current program build, if any.

fn release(&mut self)

Release all components.