Struct ocl::Program

pub struct Program(_);

A program from which kernels can be created from.

To use with multiple devices, create manually with ::from_parts().

Destruction

Handled automatically. Feel free to store, clone, and share among threads as you please.

Methods

impl Program

pub fn builder<'b>() -> ProgramBuilder<'b>

Returns a new ProgramBuilder.

pub fn with_source(
context: &ContextCore,
src_strings: &[CString],
devices: Option<&[Device]>,
cmplr_opts: &CString
) -> OclResult<Program>

Returns a new program built from pre-created build components and device list.

Prefer ::builder to create a new Program.

pub fn with_binary(
context: &ContextCore,
devices: &[Device],
binaries: &[&[u8]],
cmplr_opts: &CString
) -> OclResult<Program>

Returns a new program built from pre-created build components and device list.

Prefer ::builder to create a new Program.

pub fn with_il(
il: &[u8],
devices: Option<&[Device]>,
cmplr_opts: &CString,
context: &ContextCore
) -> OclResult<Program>

Returns a new program built from pre-created build components and device list for programs with intermediate language byte source.

pub fn as_core(&self) -> &ProgramCore

Returns a reference to the core pointer wrapper, usable by functions in the core module.

pub fn info(&self, info_kind: ProgramInfo) -> OclCoreResult<ProgramInfoResult>

Returns info about this program.

pub fn build_info(
&self,
device: Device,
info_kind: ProgramBuildInfo
) -> OclCoreResult<ProgramBuildInfoResult>

Returns info about this program's build.

TODO: Check that device is valid.