Struct ocl::Queue

pub struct Queue { /* fields omitted */ }

A command queue which manages all actions taken on kernels, buffers, and images.

Methods

impl Queue

pub fn new(
context: &Context,
device: Device,
properties: Option<CommandQueueProperties>
) -> OclResult<Queue>

Returns a new Queue on the device specified by device.

pub fn flush(&self) -> OclResult<()>

Issues all previously queued OpenCL commands to the device.

pub fn finish(&self) -> OclResult<()>

Blocks until all commands in this queue have completed before returning.

pub fn enqueue_marker<Ewl>(&self, ewait: Option<Ewl>) -> OclResult<Event> where
Ewl: ClWaitListPtr,

Enqueues a marker command which waits for either a list of events to complete, or all previously enqueued commands to complete.

pub fn as_core(&self) -> &CommandQueueCore

Returns a reference to the core pointer wrapper, usable by functions in the core module.

pub fn context(&self) -> Context

Returns a copy of the Context associated with this queue.

pub fn device(&self) -> Device

Returns the OpenCL device associated with this queue.

pub fn device_version(&self) -> OpenclVersion

Returns the cached device version.

pub fn info(
&self,
info_kind: CommandQueueInfo
) -> OclCoreResult<CommandQueueInfoResult>

Returns info about this queue.