Struct ocl::Queue

pub struct Queue {
    // some fields omitted
}

An OpenCL command queue.

Methods

impl Queue

fn new(context: &Context, device_idx: Option<usize>) -> Queue

Returns a new Queue on the device specified by device_idx.

'device_idxrefers to a index in the list of devices generated when creatingcontext. For a list of these devices, callcontext.device_ids(). Ifdevice_idxis out of range, it will automatically 'wrap around' via a modulo operation and therefore is valid up to the limit ofusize. See the documentation forContext` for more information.

fn finish(&self)

Blocks until all commands in this queue have completed.

fn obj(&self) -> cl_command_queue

Returns the OpenCL command queue object associated with this queue.

fn context_obj(&self) -> cl_context

Returns the OpenCL context object associated with this queue.

fn device_id(&self) -> cl_device_id

Returns the OpenCL device id associated with this queue.

Not to be confused with the zero-indexed device_idx passed to ::new() when creating this queue.

fn release(&mut self)

Decrements the reference counter of the associated OpenCL command queue object.