Struct ocl::Queue

pub struct Queue { /* fields omitted */ }

A command queue which manages all actions taken on kernels, buffers, and images.

Destruction

Underlying queue object is destroyed automatically.

Methods

impl Queue

fn new(context: &Context, device: Device) -> OclResult<Queue>

Returns a new Queue on the device specified by device.

fn finish(&self)

Blocks until all commands in this queue have completed before returning.

fn core_as_ref(&self) -> &CommandQueueCore

Returns a reference to the core pointer wrapper, usable by functions in the core module.

fn context_core_as_ref(&self) -> &ContextCore

Returns a reference to the core pointer wrapper of the context associated with this queue, usable by functions in the core module.

fn device(&self) -> &Device

Returns the OpenCL device associated with this queue.

fn device_version(&self) -> OpenclVersion

Returns the cached device version.

fn info(&self, info_kind: CommandQueueInfo) -> CommandQueueInfoResult

Returns info about this queue.