Struct ocl::RwVec

pub struct RwVec<T> { /* fields omitted */ }

A locking Vec which interoperates with OpenCL events and Rust futures to provide exclusive access to data.

Calling ::read or ::write returns a future which will resolve into a RwGuard.

Platform Compatibility

Some CPU device/platform combinations have synchronization problems when accessing an RwVec from multiple threads. Known platforms with problems are 2nd and 4th gen Intel Core processors (Sandy Bridge and Haswell) with Intel OpenCL CPU drivers. Others may be likewise affected. Run the device_check.rs example to determine if your device/platform is affected. AMD platform drivers are known to work properly on the aforementioned CPUs so use those instead if possible.

Methods

impl<T> RwVec<T>

pub fn new() -> RwVec<T>

Creates and returns a new RwVec.

pub fn read(self) -> FutureGuard<Vec<T>, ReadGuard<Vec<T>>>

Returns a new FutureRwGuard which will resolve into a a RwGuard.

pub fn write(self) -> FutureGuard<Vec<T>, WriteGuard<Vec<T>>>

Returns a new FutureRwGuard which will resolve into a a RwGuard.

Important traits for &'a [u8]
impl<'a> Read for &'a [u8]impl<'a> Write for &'a mut [u8]

pub unsafe fn as_mut_slice(&self) -> &mut [T]

Returns a mutable slice into the contained Vec.

Used by buffer command builders when preparing future read and write commands.

Do not use unless you are 100% certain that there will be no other reads or writes for the entire access duration (only possible if manually manipulating the lock status).

pub fn len_stale(&self) -> usize

Returns the length of the internal Vec.

Concurrency Considerations

The returned length is to be considered immediately stale (out of date) due to the fact that other threads may be modifying it. To obtain an accurate state of the internal Vec, the containing RwVec must first be locked with ::read or ::write.

pub fn id(&self) -> usize

Returns a pointer address to the internal array, usable as a unique identifier.

Note that resizing the Vec will likely change the address. Also, the same 'id' could be reused by another RwVec created after this one is dropped.